feat: remove GUI support
This commit removes support for tht GUI (or windowed) mode. In the
future, this application will be containerized and turned into a web
service, so a desktop application with a GUI is no longer relevant.
Command line mode is still relevant as an alternative to the web service.

The code still needs to be refactored, because the architecture can be
cleaner, especially without the GUI mode.